requirements.txt
streamlit
tensorflow==2.20.0
numpy
gdown
Pillow
matplotlib
scikit-learn
streamlit-extras
